LayoutTests/platform/chromium/fast/forms/calendar-picker/calendar-picker-with-step.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../fast/js/resources/js-test-pre.js"></script>
<script src="resources/calendar-picker-common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<input type=date id=date value="2011-05-01" step="9">

<script>
description('Test if step limits available dates in calendar picker.');

var dateInput = document.getElementById('date');
window.onload = function() {
    openPicker(dateInput);
    popupWindow.addEventListener('resize', test1, false);
};

function test1() {
    debug("Step without min.");
    shouldBe('availableDatesInCurrentMonth().join(",")', '["2011-05-08", "2011-05-17", "2011-05-26"].join(",")');
    popupWindow.removeEventListener('resize', test1);

    // Close popup
    eventSender.keyDown('\x1B');

    dateInput.min = "2011-05-01";
    
    // Reopen popup
    openPicker(dateInput);
    popupWindow.addEventListener("resize", test2, false);
}

function test2() {
    debug("Step when min is set.");
    shouldBe('availableDatesInCurrentMonth().join(",")', '["2011-05-01", "2011-05-10", "2011-05-19", "2011-05-28"].join(",")');
    popupWindow.removeEventListener('resize', test2);
    testRunner.notifyDone();
}

function availableDatesInCurrentMonth() {
    return Array.prototype.map.call(popupWindow.document.querySelectorAll(".day.available:not(.not-this-month)"), function(element) {
        return element.dataset.submitValue;
    }).sort();
}


</script>
<script src="../../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
